renamed developing module
