+++ zed_depth_anything.py
@@ -122,7 +122,7 @@ def main(opt):
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
+    parser = argparse.ArgumentParser(description="depth-anything with zed camera")
     parser.add_argument(
         "--input_svo_file",
         type=str,
